cflags: Set -gz by default if supported by the compiler
The -gz flag tells the compiler to use compression on the debug sections
in the output, this can reduce the disk usage of the produced object
files significantly if there is a lot of debug information.
The build system will test if the compiler supports the option and only
add it if it does not cause a syntax error in the command line.